+++ external-crates/move/crates/move-vm-runtime/src/loader/mod.rs
@@ -3,20 +3,19 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::{
+    loader::ast::{DatatypeInfo, DatatypeTagType, Function, LoadedModule},
     logging::expect_no_verification_errors,
-    native_functions::{NativeFunction, NativeFunctions, UnboxedNativeFunction},
+    native_functions::NativeFunctions,
     session::LoadedFunctionInstantiation,
 };
 use move_binary_format::{
     errors::{verification_error, Location, PartialVMError, PartialVMResult, VMResult},
     file_format::{
-        AbilitySet, Bytecode, CompiledModule, Constant, ConstantPoolIndex,
-        EnumDefInstantiationIndex, EnumDefinitionIndex, FieldHandleIndex, FieldInstantiationIndex,
-        FunctionDefinition, FunctionDefinitionIndex, FunctionHandleIndex,
+        AbilitySet, CompiledModule, Constant, ConstantPoolIndex, EnumDefinitionIndex,
+        FieldHandleIndex, FieldInstantiationIndex, FunctionDefinitionIndex, FunctionHandleIndex,
         FunctionInstantiationIndex, SignatureIndex, SignatureToken, StructDefInstantiationIndex,
         StructDefinitionIndex, StructFieldInformation, TableIndex, TypeParameterIndex,
-        VariantHandle, VariantHandleIndex, VariantInstantiationHandle,
-        VariantInstantiationHandleIndex, VariantJumpTable, VariantTag,
+        VariantHandleIndex, VariantInstantiationHandleIndex, VariantTag,
     },
     IndexKind,
 };
@@ -40,19 +39,22 @@ use move_vm_types::{
 };
 use parking_lot::RwLock;
 use std::{
-    collections::{btree_map::Entry, BTreeMap, BTreeSet, HashMap},
+    collections::{BTreeMap, BTreeSet, HashMap},
     fmt::Debug,
     hash::Hash,
     sync::Arc,
 };
 use tracing::error;
 
+pub mod ast;
+pub mod translate;
+
 // A simple cache that offers both a HashMap and a Vector lookup.
 // Values are forced into a `Arc` so they can be used from multiple thread.
 // Access to this cache is always under a `RwLock`.
-struct BinaryCache<K, V> {
-    id_map: HashMap<K, usize>,
-    binaries: Vec<Arc<V>>,
+pub(crate) struct BinaryCache<K, V> {
+    pub(crate) id_map: HashMap<K, usize>,
+    pub(crate) binaries: Vec<Arc<V>>,
 }
 
 #[derive(Debug)]
@@ -124,9 +126,9 @@ pub struct ModuleCache {
 
 /// Tracks the current end point of the `ModuleCache`'s `types`s and `function`s, so that we can
 /// roll-back to that point in case of error.
-struct CacheCursor {
-    last_datatype: usize,
-    last_function: usize,
+pub(crate) struct CacheCursor {
+    pub(crate) last_datatype: usize,
+    pub(crate) last_function: usize,
 }
 
 impl ModuleCache {
@@ -405,11 +407,11 @@ impl ModuleCache {
 
         for (idx, func) in module.function_defs().iter().enumerate() {
             let findex = FunctionDefinitionIndex(idx as TableIndex);
-            let function = Function::new(natives, findex, func, module);
+            let function = translate::function(natives, findex, func, module);
             self.functions.push(Arc::new(function));
         }
 
-        let loaded_module = LoadedModule::new(cursor, link_context, storage_id, module, self)?;
+        let loaded_module = translate::module(cursor, link_context, storage_id, module, self)?;
         self.loaded_modules
             .insert((link_context, runtime_id), loaded_module)
     }
@@ -1672,669 +1674,6 @@ impl<'a> Resolver<'a> {
     // get the loader
     pub(crate) fn loader(&self) -> &Loader {
         self.loader
-    }
-}
-
-// A LoadedModule is very similar to a CompiledModule but data is "transformed" to a representation
-// more appropriate to execution.
-// When code executes indexes in instructions are resolved against those runtime structure
-// so that any data needed for execution is immediately available
-#[derive(Debug)]
-pub(crate) struct LoadedModule {
-    #[allow(dead_code)]
-    id: ModuleId,
-
-    //
-    // types as indexes into the Loader type list
-    //
-    #[allow(dead_code)]
-    type_refs: Vec<CachedTypeIndex>,
-
-    // struct references carry the index into the global vector of types.
-    // That is effectively an indirection over the ref table:
-    // the instruction carries an index into this table which contains the index into the
-    // glabal table of types. No instantiation of generic types is saved into the global table.
-    structs: Vec<StructDef>,
-    // materialized instantiations, whether partial or not
-    struct_instantiations: Vec<StructInstantiation>,
-
-    // enum references carry the index into the global vector of types.
-    // That is effectively an indirection over the ref table:
-    // the instruction carries an index into this table which contains the index into the
-    // glabal table of types. No instantiation of generic types is saved into the global table.
-    // Note that variants are not carried in the global table as these should stay in sync with the
-    // enum type.
-    enums: Vec<EnumDef>,
-    // materialized instantiations
-    enum_instantiations: Vec<EnumInstantiation>,
-
-    variant_handles: Vec<VariantHandle>,
-    variant_instantiation_handles: Vec<VariantInstantiationHandle>,
-
-    // functions as indexes into the Loader function list
-    // That is effectively an indirection over the ref table:
-    // the instruction carries an index into this table which contains the index into the
-    // glabal table of functions. No instantiation of generic functions is saved into
-    // the global table.
-    function_refs: Vec<usize>,
-    // materialized instantiations, whether partial or not
-    function_instantiations: Vec<FunctionInstantiation>,
-
-    // fields as a pair of index, first to the type, second to the field position in that type
-    field_handles: Vec<FieldHandle>,
-    // materialized instantiations, whether partial or not
-    field_instantiations: Vec<FieldInstantiation>,
-
-    // function name to index into the Loader function list.
-    // This allows a direct access from function name to `Function`
-    function_map: HashMap<Identifier, usize>,
-
-    // a map of single-token signature indices to type.
-    // Single-token signatures are usually indexed by the `SignatureIndex` in bytecode. For example,
-    // `VecMutBorrow(SignatureIndex)`, the `SignatureIndex` maps to a single `SignatureToken`, and
-    // hence, a single type.
-    single_signature_token_map: BTreeMap<SignatureIndex, Type>,
-
-    // a map from signatures in instantiations to the `Vec<Type>` that reperesent it.
-    instantiation_signatures: BTreeMap<SignatureIndex, Vec<Type>>,
-}
-
-impl LoadedModule {
-    fn new(
-        cursor: &CacheCursor,
-        link_context: AccountAddress,
-        storage_id: ModuleId,
-        module: &CompiledModule,
-        cache: &ModuleCache,
-    ) -> Result<Self, PartialVMError> {
-        let self_id = module.self_id();
-
-        let mut instantiation_signatures: BTreeMap<SignatureIndex, Vec<Type>> = BTreeMap::new();
-        // helper to build the sparse signature vector
-        fn cache_signatures(
-            instantiation_signatures: &mut BTreeMap<SignatureIndex, Vec<Type>>,
-            module: &CompiledModule,
-            instantiation_idx: SignatureIndex,
-            cache: &ModuleCache,
-        ) -> Result<(), PartialVMError> {
-            if let Entry::Vacant(e) = instantiation_signatures.entry(instantiation_idx) {
-                let instantiation = module
-                    .signature_at(instantiation_idx)
-                    .0
-                    .iter()
-                    .map(|ty| cache.make_type(module, ty))
-                    .collect::<Result<Vec<_>, _>>()?;
-                e.insert(instantiation);
-            }
-            Ok(())
-        }
-
-        let mut type_refs = vec![];
-        let mut structs = vec![];
-        let mut struct_instantiations = vec![];
-        let mut enums = vec![];
-        let mut enum_instantiations = vec![];
-        let mut function_refs = vec![];
-        let mut function_instantiations = vec![];
-        let mut field_handles = vec![];
-        let mut field_instantiations: Vec<FieldInstantiation> = vec![];
-        let mut function_map = HashMap::new();
-        let mut single_signature_token_map = BTreeMap::new();
-
-        for datatype_handle in module.datatype_handles() {
-            let struct_name = module.identifier_at(datatype_handle.name);
-            let module_handle = module.module_handle_at(datatype_handle.module);
-            let runtime_id = module.module_id_for_handle(module_handle);
-            type_refs.push(cache.resolve_type_by_name(struct_name, &runtime_id)?.0);
-        }
-
-        for struct_def in module.struct_defs() {
-            let idx = type_refs[struct_def.struct_handle.0 as usize];
-            let field_count = cache.datatypes.binaries[idx.0].get_struct()?.fields.len() as u16;
-            structs.push(StructDef { field_count, idx });
-        }
-
-        for struct_inst in module.struct_instantiations() {
-            let def = struct_inst.def.0 as usize;
-            let struct_def = &structs[def];
-            let field_count = struct_def.field_count;
-
-            let instantiation_idx = struct_inst.type_parameters;
-            cache_signatures(
-                &mut instantiation_signatures,
-                module,
-                instantiation_idx,
-                cache,
-            )?;
-            struct_instantiations.push(StructInstantiation {
-                field_count,
-                def: struct_def.idx,
-                instantiation_idx,
-            });
-        }
-
-        for enum_def in module.enum_defs() {
-            let idx = type_refs[enum_def.enum_handle.0 as usize];
-            let enum_type = cache.datatypes.binaries[idx.0].get_enum()?;
-            let variant_count = enum_type.variants.len() as u16;
-            let variants = enum_type
-                .variants
-                .iter()
-                .enumerate()
-                .map(|(tag, variant_type)| VariantDef {
-                    tag: tag as u16,
-                    field_count: variant_type.fields.len() as u16,
-                    field_types: variant_type.fields.clone(),
-                })
-                .collect();
-            enums.push(EnumDef {
-                variant_count,
-                variants,
-                idx,
-            });
-        }
-
-        for enum_inst in module.enum_instantiations() {
-            let def = enum_inst.def.0 as usize;
-            let enum_def = &enums[def];
-            let variant_count_map = enum_def.variants.iter().map(|v| v.field_count).collect();
-            let instantiation_idx = enum_inst.type_parameters;
-            cache_signatures(
-                &mut instantiation_signatures,
-                module,
-                instantiation_idx,
-                cache,
-            )?;
-
-            enum_instantiations.push(EnumInstantiation {
-                variant_count_map,
-                def: enum_def.idx,
-                instantiation_idx,
-            });
-        }
-
-        for func_handle in module.function_handles() {
-            let func_name = module.identifier_at(func_handle.name);
-            let module_handle = module.module_handle_at(func_handle.module);
-            let runtime_id = module.module_id_for_handle(module_handle);
-            if runtime_id == self_id {
-                // module has not been published yet, loop through the functions
-                for (idx, function) in cache.functions.iter().enumerate().rev() {
-                    if idx < cursor.last_function {
-                        return Err(PartialVMError::new(StatusCode::FUNCTION_RESOLUTION_FAILURE)
-                            .with_message(format!(
-                                "Cannot find {:?}::{:?} in publishing module",
-                                runtime_id, func_name
-                            )));
-                    }
-                    if function.name.as_ident_str() == func_name {
-                        function_refs.push(idx);
-                        break;
-                    }
-                }
-            } else {
-                function_refs.push(cache.resolve_function_by_name(
-                    func_name,
-                    &runtime_id,
-                    link_context,
-                )?);
-            }
-        }
-
-        for func_def in module.function_defs() {
-            let idx = function_refs[func_def.function.0 as usize];
-            let name = module.identifier_at(module.function_handle_at(func_def.function).name);
-            function_map.insert(name.to_owned(), idx);
-
-            if let Some(code_unit) = &func_def.code {
-                for bc in &code_unit.code {
-                    match bc {
-                        Bytecode::VecPack(si, _)
-                        | Bytecode::VecLen(si)
-                        | Bytecode::VecImmBorrow(si)
-                        | Bytecode::VecMutBorrow(si)
-                        | Bytecode::VecPushBack(si)
-                        | Bytecode::VecPopBack(si)
-                        | Bytecode::VecUnpack(si, _)
-                        | Bytecode::VecSwap(si) => {
-                            if !single_signature_token_map.contains_key(si) {
-                                let ty = match module.signature_at(*si).0.first() {
-                                    None => {
-                                        return Err(PartialVMError::new(
-                                            StatusCode::VERIFIER_INVARIANT_VIOLATION,
-                                        )
-                                        .with_message(
-                                            "the type argument for vector-related bytecode \
-                                                expects one and only one signature token"
-                                                .to_owned(),
-                                        ));
-                                    }
-                                    Some(sig_token) => sig_token,
-                                };
-                                single_signature_token_map
-                                    .insert(*si, cache.make_type(module, ty)?);
-                            }
-                        }
-                        _ => {}
-                    }
-                }
-            }
-        }
-
-        for func_inst in module.function_instantiations() {
-            let handle = function_refs[func_inst.handle.0 as usize];
-
-            let instantiation_idx = func_inst.type_parameters;
-            cache_signatures(
-                &mut instantiation_signatures,
-                module,
-                instantiation_idx,
-                cache,
-            )?;
-            function_instantiations.push(FunctionInstantiation {
-                handle,
-                instantiation_idx,
-            });
-        }
-
-        for f_handle in module.field_handles() {
-            let def_idx = f_handle.owner;
-            let owner = structs[def_idx.0 as usize].idx;
-            let offset = f_handle.field as usize;
-            field_handles.push(FieldHandle { offset, owner });
-        }
-
-        for f_inst in module.field_instantiations() {
-            let fh_idx = f_inst.handle;
-            let owner = field_handles[fh_idx.0 as usize].owner;
-            let offset = field_handles[fh_idx.0 as usize].offset;
-
-            field_instantiations.push(FieldInstantiation { offset, owner });
-        }
-
-        Ok(Self {
-            id: storage_id,
-            type_refs,
-            structs,
-            struct_instantiations,
-            enums,
-            enum_instantiations,
-            function_refs,
-            function_instantiations,
-            field_handles,
-            field_instantiations,
-            function_map,
-            single_signature_token_map,
-            instantiation_signatures,
-            variant_handles: module.variant_handles().to_vec(),
-            variant_instantiation_handles: module.variant_instantiation_handles().to_vec(),
-        })
-    }
-
-    fn struct_at(&self, idx: StructDefinitionIndex) -> CachedTypeIndex {
-        self.structs[idx.0 as usize].idx
-    }
-
-    fn struct_instantiation_at(&self, idx: u16) -> &StructInstantiation {
-        &self.struct_instantiations[idx as usize]
-    }
-
-    fn function_at(&self, idx: u16) -> usize {
-        self.function_refs[idx as usize]
-    }
-
-    fn function_instantiation_at(&self, idx: u16) -> &FunctionInstantiation {
-        &self.function_instantiations[idx as usize]
-    }
-
-    fn field_count(&self, idx: u16) -> u16 {
-        self.structs[idx as usize].field_count
-    }
-
-    fn field_instantiation_count(&self, idx: u16) -> u16 {
-        self.struct_instantiations[idx as usize].field_count
-    }
-
-    fn field_offset(&self, idx: FieldHandleIndex) -> usize {
-        self.field_handles[idx.0 as usize].offset
-    }
-
-    fn field_instantiation_offset(&self, idx: FieldInstantiationIndex) -> usize {
-        self.field_instantiations[idx.0 as usize].offset
-    }
-
-    fn single_type_at(&self, idx: SignatureIndex) -> &Type {
-        self.single_signature_token_map.get(&idx).unwrap()
-    }
-
-    fn instantiation_signature_at(
-        &self,
-        idx: SignatureIndex,
-    ) -> Result<&Vec<Type>, PartialVMError> {
-        self.instantiation_signatures.get(&idx).ok_or_else(|| {
-            PartialVMError::new(StatusCode::UNKNOWN_INVARIANT_VIOLATION_ERROR)
-                .with_message("Instantiation signature not found".to_string())
-        })
-    }
-
-    fn enum_at(&self, idx: EnumDefinitionIndex) -> CachedTypeIndex {
-        self.enums[idx.0 as usize].idx
-    }
-
-    fn enum_instantiation_at(&self, idx: EnumDefInstantiationIndex) -> &EnumInstantiation {
-        &self.enum_instantiations[idx.0 as usize]
-    }
-
-    fn variant_at(&self, vidx: VariantHandleIndex) -> &VariantDef {
-        let variant_handle = &self.variant_handles[vidx.0 as usize];
-        let enum_def = &self.enums[variant_handle.enum_def.0 as usize];
-        &enum_def.variants[variant_handle.variant as usize]
-    }
-
-    fn variant_handle_at(&self, vidx: VariantHandleIndex) -> &VariantHandle {
-        &self.variant_handles[vidx.0 as usize]
-    }
-
-    fn variant_field_count(&self, vidx: VariantHandleIndex) -> (u16, VariantTag) {
-        let variant = self.variant_at(vidx);
-        (variant.field_count, variant.tag)
-    }
-
-    fn variant_instantiation_handle_at(
-        &self,
-        vidx: VariantInstantiationHandleIndex,
-    ) -> &VariantInstantiationHandle {
-        &self.variant_instantiation_handles[vidx.0 as usize]
-    }
-
-    fn variant_instantiantiation_field_count_and_tag(
-        &self,
-        vidx: VariantInstantiationHandleIndex,
-    ) -> (u16, VariantTag) {
-        let handle = self.variant_instantiation_handle_at(vidx);
-        let enum_inst = &self.enum_instantiations[handle.enum_def.0 as usize];
-        (
-            enum_inst.variant_count_map[handle.variant as usize],
-            handle.variant,
-        )
-    }
-}
-
-// A runtime function
-// #[derive(Debug)]
-// https://github.com/rust-lang/rust/issues/70263
-pub(crate) struct Function {
-    #[allow(unused)]
-    file_format_version: u32,
-    index: FunctionDefinitionIndex,
-    code: Vec<Bytecode>,
-    parameters: SignatureIndex,
-    return_: SignatureIndex,
-    type_parameters: Vec<AbilitySet>,
-    native: Option<NativeFunction>,
-    def_is_native: bool,
-    module: ModuleId,
-    name: Identifier,
-    parameters_len: usize,
-    locals_len: usize,
-    return_len: usize,
-    jump_tables: Vec<VariantJumpTable>,
-}
-
-impl Function {
-    fn new(
-        natives: &NativeFunctions,
-        index: FunctionDefinitionIndex,
-        def: &FunctionDefinition,
-        module: &CompiledModule,
-    ) -> Self {
-        let handle = module.function_handle_at(def.function);
-        let name = module.identifier_at(handle.name).to_owned();
-        let module_id = module.self_id();
-        let (native, def_is_native) = if def.is_native() {
-            (
-                natives.resolve(
-                    module_id.address(),
-                    module_id.name().as_str(),
-                    name.as_str(),
-                ),
-                true,
-            )
-        } else {
-            (None, false)
-        };
-        let parameters = handle.parameters;
-        let parameters_len = module.signature_at(parameters).0.len();
-        // Native functions do not have a code unit
-        let (code, locals_len, jump_tables) = match &def.code {
-            Some(code) => (
-                code.code.clone(),
-                parameters_len + module.signature_at(code.locals).0.len(),
-                code.jump_tables.clone(),
-            ),
-            None => (vec![], 0, vec![]),
-        };
-        let return_ = handle.return_;
-        let return_len = module.signature_at(return_).0.len();
-        let type_parameters = handle.type_parameters.clone();
-        Self {
-            file_format_version: module.version(),
-            index,
-            code,
-            parameters,
-            return_,
-            type_parameters,
-            native,
-            def_is_native,
-            module: module_id,
-            name,
-            parameters_len,
-            locals_len,
-            return_len,
-            jump_tables,
-        }
-    }
-
-    #[allow(unused)]
-    pub(crate) fn file_format_version(&self) -> u32 {
-        self.file_format_version
-    }
-
-    pub(crate) fn module_id(&self) -> &ModuleId {
-        &self.module
-    }
-
-    pub(crate) fn index(&self) -> FunctionDefinitionIndex {
-        self.index
-    }
-
-    pub(crate) fn get_resolver<'a>(
-        &self,
-        link_context: AccountAddress,
-        loader: &'a Loader,
-    ) -> Resolver<'a> {
-        let (compiled, loaded) = loader.get_module(link_context, &self.module);
-        Resolver::for_module(loader, compiled, loaded)
-    }
-
-    pub(crate) fn local_count(&self) -> usize {
-        self.locals_len
-    }
-
-    pub(crate) fn arg_count(&self) -> usize {
-        self.parameters_len
-    }
-
-    pub(crate) fn return_type_count(&self) -> usize {
-        self.return_len
-    }
-
-    pub(crate) fn name(&self) -> &str {
-        self.name.as_str()
-    }
-
-    pub(crate) fn code(&self) -> &[Bytecode] {
-        &self.code
-    }
-
-    pub(crate) fn jump_tables(&self) -> &[VariantJumpTable] {
-        &self.jump_tables
-    }
-
-    pub(crate) fn type_parameters(&self) -> &[AbilitySet] {
-        &self.type_parameters
-    }
-
-    pub(crate) fn pretty_string(&self) -> String {
-        let id = &self.module;
-        format!(
-            "0x{}::{}::{}",
-            id.address(),
-            id.name().as_str(),
-            self.name.as_str()
-        )
-    }
-
-    #[cfg(any(debug_assertions, feature = "debugging"))]
-    pub(crate) fn pretty_short_string(&self) -> String {
-        let id = &self.module;
-        format!(
-            "0x{}::{}::{}",
-            id.address().short_str_lossless(),
-            id.name().as_str(),
-            self.name.as_str()
-        )
-    }
-
-    pub(crate) fn is_native(&self) -> bool {
-        self.def_is_native
-    }
-
-    pub(crate) fn get_native(&self) -> PartialVMResult<&UnboxedNativeFunction> {
-        if cfg!(feature = "lazy_natives") {
-            // If lazy_natives is configured, this is a MISSING_DEPENDENCY error, as we skip
-            // checking those at module loading time.
-            self.native.as_deref().ok_or_else(|| {
-                PartialVMError::new(StatusCode::MISSING_DEPENDENCY)
-                    .with_message(format!("Missing Native Function `{}`", self.name))
-            })
-        } else {
-            // Otherwise this error should not happen, hence UNREACHABLE
-            self.native.as_deref().ok_or_else(|| {
-                PartialVMError::new(StatusCode::UNREACHABLE)
-                    .with_message("Missing Native Function".to_string())
-            })
-        }
-    }
-}
-
-//
-// Internal structures that are saved at the proper index in the proper tables to access
-// execution information (interpreter).
-// The following structs are internal to the loader and never exposed out.
-// The `Loader` will create those struct and the proper table when loading a module.
-// The `Resolver` uses those structs to return information to the `Interpreter`.
-//
-
-// A function instantiation.
-#[derive(Debug)]
-struct FunctionInstantiation {
-    // index to `ModuleCache::functions` global table
-    handle: usize,
-    instantiation_idx: SignatureIndex,
-}
-
-#[derive(Debug)]
-struct StructDef {
-    // struct field count
-    field_count: u16,
-    // `ModuelCache::structs` global table index
-    idx: CachedTypeIndex,
-}
-
-#[derive(Debug)]
-struct StructInstantiation {
-    // struct field count
-    field_count: u16,
-    // `ModuleCache::structs` global table index. It is the generic type.
-    def: CachedTypeIndex,
-    instantiation_idx: SignatureIndex,
-}
-
-// A field handle. The offset is the only used information when operating on a field
-#[derive(Debug)]
-struct FieldHandle {
-    offset: usize,
-    // `ModuelCache::structs` global table index. It is the generic type.
-    owner: CachedTypeIndex,
-}
-
-// A field instantiation. The offset is the only used information when operating on a field
-#[derive(Debug)]
-struct FieldInstantiation {
-    offset: usize,
-    // `ModuleCache::structs` global table index. It is the generic type.
-    #[allow(unused)]
-    owner: CachedTypeIndex,
-}
-
-#[derive(Debug)]
-struct EnumDef {
-    // enum variant count
-    #[allow(unused)]
-    variant_count: u16,
-    variants: Vec<VariantDef>,
-    // `ModuelCache::types` global table index
-    idx: CachedTypeIndex,
-}
-
-#[derive(Debug)]
-struct EnumInstantiation {
-    // enum variant count
-    variant_count_map: Vec<u16>,
-    // `ModuelCache::types` global table index
-    def: CachedTypeIndex,
-    instantiation_idx: SignatureIndex,
-}
-
-#[derive(Debug)]
-struct VariantDef {
-    #[allow(unused)]
-    tag: u16,
-    field_count: u16,
-    #[allow(unused)]
-    field_types: Vec<Type>,
-}
-
-//
-// Cache for data associated to a Struct, used for de/serialization and more
-//
-
-struct DatatypeInfo {
-    runtime_tag: Option<StructTag>,
-    defining_tag: Option<StructTag>,
-    layout: Option<R::MoveDatatypeLayout>,
-    annotated_layout: Option<A::MoveDatatypeLayout>,
-    node_count: Option<u64>,
-    annotated_node_count: Option<u64>,
-}
-
-#[derive(Copy, Clone, PartialEq, Eq)]
-enum DatatypeTagType {
-    Runtime,
-    Defining,
-}
-
-impl DatatypeInfo {
-    fn new() -> Self {
-        Self {
-            runtime_tag: None,
-            defining_tag: None,
-            layout: None,
-            annotated_layout: None,
-            node_count: None,
-            annotated_node_count: None,
-        }
     }
 }
